+++ CHANGELOG.md
@@ -21,7 +21,7 @@ The current status of the libraries at the time of the release is as follows:
 - add support for ICP Index canister to library `@dfinity/ledger-icp`. New `IndexCanister` functions: `accountBalance` and `getTransactions`
 - expose few types - notably `BlockHeight` - for library `@dfinity/ledger-icp`
 - support new fields from swap canister response types: `min_direct_participation_icp_e8s`, `max_direct_participation_icp_e8s` and `neurons_fund_participation`
-- support new fields in the `CreateServiceNervousSystem` proposal action `maximum_direct_participation_icp`, `minimum_direct_participation_icp` and `neurons_fund_participation`
+- support new fields in the `CreateServiceNervousSystem` proposal action: `maximum_direct_participation_icp`, `minimum_direct_participation_icp` and `neurons_fund_participation`
 - support new filter field `omit_large_fields` in `list_proposals`
 - add support for `retrieve_btc_with_approval` in `@dfinity/ckbtc`
 - new utility to convert seconds to a human-readable duration